+++ SourceCode/Database/dbo/Tables/Meeting.sql
@@ -1,16 +1,19 @@
 ﻿CREATE TABLE [dbo].[Meeting] (
     [Id]               INT           IDENTITY (1, 1) NOT NULL,
     [OrganiserGroupId] INT           NOT NULL,
-    [PlaceName]        NVARCHAR (50) NOT NULL,
+    [CityName]         NVARCHAR (50) NOT NULL,
+    [VenueName]        NVARCHAR (50) NULL,
     [Name]             NVARCHAR (50) NOT NULL,
     [StartDate]        DATETIME      NOT NULL,
     [EndDate]          DATETIME      NOT NULL,
     [Status]           INT           NOT NULL,
-    [IsFremo]          BIT           CONSTRAINT [DF_Meeting_IsFremoMeeting] DEFAULT ((0)) NOT NULL,
-    [Accomodation]    NVARCHAR(MAX) NULL, 
-    [Details]      NVARCHAR(MAX) NULL, 
+    [GroupDomainId]    INT           NULL,
+    [Accomodation]     NVARCHAR(MAX) NULL, 
+    [Details]          NVARCHAR(MAX) NULL, 
     CONSTRAINT [PK_Meeting] PRIMARY KEY CLUSTERED ([Id] ASC),
-    CONSTRAINT [FK_Meeting_Group] FOREIGN KEY ([OrganiserGroupId]) REFERENCES [dbo].[Group] ([Id])
+    CONSTRAINT [FK_Meeting_Group] FOREIGN KEY ([OrganiserGroupId]) REFERENCES [dbo].[Group] ([Id]),
+    CONSTRAINT [FK_Meeting_GroupDomain] FOREIGN KEY ([GroupDomainId]) REFERENCES [dbo].[GroupDomain] ([Id])
+
 );
 GO
 CREATE TRIGGER [DeleteMeeting] ON [Meeting] INSTEAD OF DELETE